Moving gRPC stub-with-auth helpers out of bigtable package.
Also changing the metadata plugin to take credentials
and user agent directly instead of inferring those values
from a client (this is because the Bigtable client is
unique in it's attributes).
