+++ gcloud/test__helpers.py
@@ -800,6 +800,120 @@ class Test__name_from_project_path(unittest.TestCase):
         self.assertEqual(name, self.THING_NAME)
 
 
+class TestMetadataPlugin(unittest.TestCase):
+
+    def _getTargetClass(self):
+        from gcloud._helpers import MetadataPlugin
+        return MetadataPlugin
+
+    def _makeOne(self, *args, **kwargs):
+        return self._getTargetClass()(*args, **kwargs)
+
+    def test_constructor(self):
+        credentials = object()
+        user_agent = object()
+        plugin = self._makeOne(credentials, user_agent)
+        self.assertIs(plugin._credentials, credentials)
+        self.assertIs(plugin._user_agent, user_agent)
+
+    def test___call__(self):
+        access_token_expected = 'FOOBARBAZ'
+        credentials = _Credentials(access_token=access_token_expected)
+        user_agent = 'USER_AGENT'
+        callback_args = []
+
+        def callback(*args):
+            callback_args.append(args)
+
+        transformer = self._makeOne(credentials, user_agent)
+        result = transformer(None, callback)
+        cb_headers = [
+            ('Authorization', 'Bearer ' + access_token_expected),
+            ('User-agent', user_agent),
+        ]
+        self.assertEqual(result, None)
+        self.assertEqual(callback_args, [(cb_headers, None)])
+        self.assertEqual(len(credentials._tokens), 1)
+
+
+class Test_make_stub(unittest.TestCase):
+
+    def _callFUT(self, *args, **kwargs):
+        from gcloud._helpers import make_stub
+        return make_stub(*args, **kwargs)
+
+    def test_it(self):
+        from gcloud._testing import _Monkey
+        from gcloud import _helpers as MUT
+
+        mock_result = object()
+        stub_inputs = []
+
+        SSL_CREDS = object()
+        METADATA_CREDS = object()
+        COMPOSITE_CREDS = object()
+        CHANNEL = object()
+
+        class _ImplementationsModule(object):
+
+            def __init__(self):
+                self.ssl_channel_credentials_args = None
+                self.metadata_call_credentials_args = None
+                self.composite_channel_credentials_args = None
+                self.secure_channel_args = None
+
+            def ssl_channel_credentials(self, *args):
+                self.ssl_channel_credentials_args = args
+                return SSL_CREDS
+
+            def metadata_call_credentials(self, *args, **kwargs):
+                self.metadata_call_credentials_args = (args, kwargs)
+                return METADATA_CREDS
+
+            def composite_channel_credentials(self, *args):
+                self.composite_channel_credentials_args = args
+                return COMPOSITE_CREDS
+
+            def secure_channel(self, *args):
+                self.secure_channel_args = args
+                return CHANNEL
+
+        implementations_mod = _ImplementationsModule()
+
+        def mock_stub_factory(channel):
+            stub_inputs.append(channel)
+            return mock_result
+
+        metadata_plugin = object()
+        plugin_args = []
+
+        def mock_plugin(*args):
+            plugin_args.append(args)
+            return metadata_plugin
+
+        host = 'HOST'
+        port = 1025
+        credentials = object()
+        user_agent = 'USER_AGENT'
+        with _Monkey(MUT, implementations=implementations_mod,
+                     MetadataPlugin=mock_plugin):
+            result = self._callFUT(credentials, user_agent,
+                                   mock_stub_factory, host, port)
+
+        self.assertTrue(result is mock_result)
+        self.assertEqual(stub_inputs, [CHANNEL])
+        self.assertEqual(plugin_args, [(credentials, user_agent)])
+        self.assertEqual(implementations_mod.ssl_channel_credentials_args,
+                         (None, None, None))
+        self.assertEqual(implementations_mod.metadata_call_credentials_args,
+                         ((metadata_plugin,), {'name': 'google_creds'}))
+        self.assertEqual(
+            implementations_mod.composite_channel_credentials_args,
+            (SSL_CREDS, METADATA_CREDS))
+        self.assertEqual(implementations_mod.secure_channel_args,
+                         (host, port, COMPOSITE_CREDS))
+
+
 class _AppIdentity(object):
 
     def __init__(self, app_id):
@@ -852,3 +966,17 @@ class _TimeoutHTTPConnection(_BaseHTTPConnection):
     def getresponse(self):
         import socket
         raise socket.timeout('timed out')
+
+
+class _Credentials(object):
+
+    def __init__(self, access_token=None):
+        self._access_token = access_token
+        self._tokens = []
+
+    def get_access_token(self):
+        from oauth2client.client import AccessTokenInfo
+        token = AccessTokenInfo(access_token=self._access_token,
+                                expires_in=None)
+        self._tokens.append(token)
+        return token
